Refactored to fix circular imports.
Also moved a bunch of objects around to where they make the most sense
thereby reducing the mental effort required to understand each module.
